+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Updated counter metrics to incorporate the suffix `_total`. If you are using a custom dashboard to monitor Teku metrics, you might need to update the metrics manually when breaking changes are introduced. For more information, see [Update metrics](../../how-to/monitor/update-metrics.md).
 
 ### Additions and Improvements
+- Added a state pruner that can limit the number of finalized states stored when running an archive node.
 
 - Updated bootnodes for Sepolia network.